ml-model-api/requirements.txt
# python==3.10.14
Flask==3.0.2
Flask-CORS==4.0.0
numpy==1.26.4
scikit-learn==1.4.2
textblob==0.17.1
newspaper3k==0.2.8
joblib==1.3.2
python-dotenv==1.0.1
requests==2.31.0
lxml==4.9.4
gunicorn==21.2.0


google-generativeai
